getting-started/requirements.txt
mlflow==1.27.0
python-dotenv==0.21.0
pyyaml==6.0
